Fields validation, delete and close warning messages

+++ StockApp/StockApp/Products.cs
@@ -20,43 +20,61 @@ namespace StockApp
 
         private void Products_Load(object sender, EventArgs e)
         {
-            productStatusComboBox.SelectedIndex = 1;
+            productStatusComboBox.SelectedIndex = -1;
+            productNameTextBox.Focus();
+
             LoadData();
         }
 
         private void addButton_Click(object sender, EventArgs e)
         {
-            SqlConnection connection = new SqlConnection("Data Source = DEV-VM\\SQLSERVER2019;Initial Catalog = Stock;Integrated Security = True");
-            connection.Open();
-
-            var sqlQuery = "";
-
-            if (IfProductExists(connection, productCodeTextBox.Text))
+            if (Validation())
             {
-                //update record
-                sqlQuery = @"UPDATE [dbo].[Products] 
+                //SqlConnection connection = new SqlConnection("Data Source = DEV-VM\\SQLSERVER2019;Initial Catalog = Stock;Integrated Security = True");
+                SqlConnection connection = DBConnection.GetConnection();
+                connection.Open();
+
+                var sqlQuery = "";
+
+                if (IfProductExists(connection, productCodeTextBox.Text))
+                {
+                    //update record
+                    sqlQuery = @"UPDATE [dbo].[Products] 
                             SET [ProductName] = '" + productNameTextBox.Text + "', [ProductStatus] = " + productStatusComboBox.SelectedIndex + " WHERE [ProductCode] = " + productCodeTextBox.Text;
-            }
-            else
-            {
-                // insert new record - not exists
-                sqlQuery = @"INSERT INTO [dbo].[Products] ([ProductName], [ProductStatus])
+                }
+                else
+                {
+                    // insert new record - not exists
+                    sqlQuery = @"INSERT INTO [dbo].[Products] ([ProductName], [ProductStatus])
                             VALUES ('" + productNameTextBox.Text + "', '" + productStatusComboBox.SelectedIndex + "')";
+                }
+
+                SqlCommand cmd = new SqlCommand(sqlQuery, connection);
+                cmd.ExecuteNonQuery();
+                connection.Close();
+
+                // Load data to DataGridView
+                LoadData();
+
+                ClearTextBoxes(); 
             }
+        }
 
-            SqlCommand cmd = new SqlCommand(sqlQuery, connection);
-            cmd.ExecuteNonQuery();
-            connection.Close();
-
-            // Load data to DataGridView
-            LoadData();
+        private void ClearTextBoxes()
+        {
+            productCodeTextBox.Clear();
+            productNameTextBox.Clear();
+            productStatusComboBox.SelectedIndex = -1;
+            productStatusComboBox.Text = "";
+            addButton.Text = "Add";
+            productNameTextBox.Focus();
         }
 
         private bool IfProductExists(SqlConnection connection ,string productCode)
         {
             if(productCode =="")
             {
-                MessageBox.Show("Please select a Product.");
+                //MessageBox.Show("Please select a Product.");
                 return false;
             }
             else
@@ -73,7 +91,7 @@ namespace StockApp
         }
         private void LoadData()
         {
-            SqlConnection connection = new SqlConnection("Data Source = DEV-VM\\SQLSERVER2019;Initial Catalog = Stock;Integrated Security = True");
+            SqlConnection connection = DBConnection.GetConnection();
 
             SqlDataAdapter sqlDataAdapter = new SqlDataAdapter("SELECT * FROM [dbo].[Products]", connection);
             DataTable dataTable = new DataTable();
@@ -97,30 +115,84 @@ namespace StockApp
             productCodeTextBox.Text = productsDataGridView.SelectedRows[0].Cells[0].Value.ToString();
             productNameTextBox.Text = productsDataGridView.SelectedRows[0].Cells[1].Value.ToString();
             productStatusComboBox.Text = productsDataGridView.SelectedRows[0].Cells[2].Value.ToString();
+            //productStatusComboBox.SelectedIndex = int.Parse(productsDataGridView.SelectedRows[0].Cells[2].Value.ToString()); //productStatusComboBox.SelectedIndex;
+
+            //MessageBox.Show(productsDataGridView.SelectedRows[0].Cells[2].Value.ToString());//true
+            //MessageBox.Show(productsDataGridView.SelectedRows[0].Cells[2].ToString()); // row index....
+            //MessageBox.Show(productsDataGridView.SelectedRows[0].Cells[2].Selected.ToString() );//true
+            //MessageBox.Show(productsDataGridView.SelectedRows[0].Cells[2].State.ToString()); //none
+            //MessageBox.Show(productsDataGridView.SelectedRows[0].Cells[2].);
+
+            addButton.Text = "Update";
         }
 
         private void deleteButton_Click(object sender, EventArgs e)
         {
-            SqlConnection connection = new SqlConnection("Data Source = DEV-VM\\SQLSERVER2019;Initial Catalog = Stock;Integrated Security = True");
-
-            var sqlQuery = "";
-
-            if (IfProductExists(connection, productCodeTextBox.Text))
+            DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete record?", "Message", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
             {
-                //delete record
-                connection.Open();
-                sqlQuery = @"DELETE FROM [dbo].[Products] WHERE [ProductCode] = " + productCodeTextBox.Text;
+                if (Validation())
+                {
+                    SqlConnection connection = DBConnection.GetConnection();
 
-                SqlCommand cmd = new SqlCommand(sqlQuery, connection);
-                cmd.ExecuteNonQuery();
-                connection.Close();
+                    var sqlQuery = "";
+
+                    if (IfProductExists(connection, productCodeTextBox.Text))
+                    {
+                        //delete record
+                        connection.Open();
+                        sqlQuery = @"DELETE FROM [dbo].[Products] WHERE [ProductCode] = " + productCodeTextBox.Text;
+
+                        SqlCommand cmd = new SqlCommand(sqlQuery, connection);
+                        cmd.ExecuteNonQuery();
+                        connection.Close();
+                    }
+                    else
+                    {
+                        MessageBox.Show("You must select a valid Product to delete.", "Error");
+                    }
+                    // Load data to DataGridView
+                    LoadData();
+                    ClearTextBoxes();
+                } 
+            }
+        }
+
+        private void ResetButton_Click(object sender, EventArgs e)
+        {
+            ClearTextBoxes();
+        }
+
+        private bool Validation()
+        {
+            bool result = false;
+
+            errorProvider.Clear();
+
+            //if (string.IsNullOrEmpty(productCodeTextBox.Text))
+            //{
+            //    errorProvider.Clear();
+            //    errorProvider.SetError(productCodeTextBox, "Product Name Required");
+            //}
+            //else if (string.IsNullOrEmpty(productNameTextBox.Text))
+            if (string.IsNullOrEmpty(productNameTextBox.Text))
+            {
+                errorProvider.Clear();
+                errorProvider.SetError(productNameTextBox, "Product name required");
+            }
+            //else if (productStatusComboBox.SelectedIndex == -1)
+            else if (productStatusComboBox.Text == "")
+            {
+                errorProvider.Clear();
+                errorProvider.SetError(productStatusComboBox, "Pease select status");
             }
             else
-            {
-                MessageBox.Show("You must select a valid Product to delete.","Error");
-            }
-            // Load data to DataGridView
-            LoadData();
+                result = true;
+
+           //if (string.IsNullOrEmpty(productCodeTextBox.Text) && string.IsNullOrEmpty(productNameTextBox.Text) && productStatusComboBox.SelectedIndex == -1)
+           //result = false;
+           
+            return result;
         }
     }
-} // comment test for GitHub
+}
